requirements.txt
#requirements.txt

# Core
fastapi==0.111.0
uvicorn==0.30.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3

# Dashboard
streamlit==1.35.0
pandas==2.2.2
plotly==5.22.0
streamlit-autorefresh==1.0.1
